requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.9

# Email processing
imap-tools==1.12.0
google-api-python-client==2.108.0

# PDF processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.17.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2

# NLP/Classification
python-dateutil==2.8.2

# Bank statement parsing
ofxparse==0.21

# Scheduling
apscheduler==3.10.4

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
